+++ src/Data/CourseReadinessContext.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Lalalili\CourseCore\Data;
+
+final readonly class CourseReadinessContext
+{
+    public function __construct(
+        public bool $requireProduct = false,
+        public bool $requireReadyVideos = false,
+    ) {
+    }
+}